port check 054: "Show hinting filesize impact."
(issue #1388)
