+++ Lib/fontbakery/checks.py
@@ -250,70 +250,6 @@ def perform_all_fontforge_checks(fb, validation_state):
            "Hinds do not overlap.")
 
 
-def check_hinting_filesize_impact(fb, fullpath, filename):
-  fb.new_check("054", "Show hinting filesize impact")
-  # current implementation simply logs useful info
-  # but there's no fail scenario for this checker.
-  ttfautohint_missing = False
-  try:
-    import subprocess
-    statinfo = os.stat(fullpath)
-    hinted_size = statinfo.st_size
-
-    dehinted = tempfile.NamedTemporaryFile(suffix=".ttf", delete=False)
-    subprocess.call(["ttfautohint",
-                     "--dehint",
-                     fullpath,
-                     dehinted.name])
-    statinfo = os.stat(dehinted.name)
-    dehinted_size = statinfo.st_size
-    os.unlink(dehinted.name)
-
-    if dehinted_size == 0:
-      fb.skip("ttfautohint --dehint reports that"
-              " 'This font has already been processed with ttfautohint'."
-              " This is a bug in an old version of ttfautohint."
-              " You'll need to upgrade it."
-              " See https://github.com/googlefonts/fontbakery/"
-              "issues/1043#issuecomment-249035069")
-    else:
-      increase = hinted_size - dehinted_size
-      change = float(hinted_size)/dehinted_size - 1
-      change = int(change*10000)/100.0  # round to 2 decimal pts percentage
-
-      def filesize_formatting(s):
-        if s < 1024:
-          return "{} bytes".format(s)
-        elif s < 1024*1024:
-          return "{}kb".format(s/1024)
-        else:
-          return "{}Mb".format(s/(1024*1024))
-
-      hinted_size = filesize_formatting(hinted_size)
-      dehinted_size = filesize_formatting(dehinted_size)
-      increase = filesize_formatting(increase)
-
-      results_table = "Hinting filesize impact:\n\n"
-      results_table += "|  | {} |\n".format(filename)
-      results_table += "|:--- | ---:| ---:|\n"
-      results_table += "| Dehinted Size | {} |\n".format(dehinted_size)
-      results_table += "| Hinted Size | {} |\n".format(hinted_size)
-      results_table += "| Increase | {} |\n".format(increase)
-      results_table += "| Change   | {} % |\n".format(change)
-      fb.info(results_table)
-
-  except OSError:
-    # This is made very prominent with additional line breaks
-    ttfautohint_missing = True
-    fb.warning("\n\n\nttfautohint is not available!"
-               " You really MUST check the fonts with this tool."
-               " To install it, see"
-               " https://github.com/googlefonts"
-               "/gf-docs/blob/master/"
-               "ProjectChecklist.md#ttfautohint\n\n\n")
-  return ttfautohint_missing
-
-
 def check_version_format_is_correct_in_NAME_table(fb, font):
   fb.new_check("055", "Version format is correct in NAME table?")
