+++ bin/fontbakery-check-ttf.py
@@ -204,13 +204,10 @@ def fontbakery_check_ttf(config):
       checks.check_glyphset_google_vietnamese(fb, target)
       checks.check_glyphset_google_extras(fb, target)
 
-    ttfautohint_missing = checks.check_hinting_filesize_impact(fb,
-                                                               target,
-                                                               filename)
     checks.check_version_format_is_correct_in_NAME_table(fb, font)
-    checks.check_font_has_latest_ttfautohint_applied(fb,
-                                                     font,
-                                                     ttfautohint_missing)
+    #checks.check_font_has_latest_ttfautohint_applied(fb,
+    #                                                 font,
+    #                                                 ttfautohint_missing)
     checks.check_name_table_entries_do_not_contain_linebreaks(fb, font)
     checks.check_glyph_names_are_all_valid(fb, font)
     checks.check_font_has_unique_glyph_names(fb, font)
